+++ services/cookbook/core/src/main/java/family/haschka/wolkenschloss/cookbook/recipe/WrappedString.java
