+++ services/cookbook/core/src/test/java/family/haschka/wolkenschloss/cookbook/recipe/WrappedStringTest.java
